+++ miniprogram/config/env.js
@@ -2,8 +2,8 @@
 const FALLBACK_ENV = "release";
 
 const ENV_API_BASE_URL_MAP = {
-  //develop: "https://hhycanteen.iepose.cn/api/v1",
   develop: "http://127.0.0.1:8000/api/v1",
+  //develop: "https://hhycanteen.iepose.cn/api/v1",
   trial: "https://hhycanteen.iepose.cn/api/v1",
   release: "https://hhycanteen.iepose.cn/api/v1"
 };